requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-dotenv==1.0.0
feedparser==6.0.11
reportlab==4.0.4
jinja2==3.1.2
werkzeug==2.3.7
gunicorn==21.2.0
pytz==2023.3
python-dateutil==2.8.2
schedule==1.2.0
lxml==4.9.3
praw==7.7.1
Pillow==10.0.1